backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "echodesk-voice-backend"
version = "1.0.0"
description = "Python FastAPI voice backend for Echodesk AI receptionist"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["voice*", "prompts*", "telnyx*", "calendar_api*"]
